requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
pg8000==1.30.3
reportlab==4.0.7
python-dotenv==1.0.0
email-validator==2.1.0
Pillow==10.1.0
openpyxl==3.1.2
WTForms==3.1.1
